Fix image upload
Make image upload and display work. The directory images were being
uploaded to was being deleted on start and every 10 minutes. With no
directory to write the files to, the uploads were failing. This was
probably due to changes in Sails since original implementation.

This 'fix' isn't good, but at least it works.